EES-5733 fix table download and tidying
