+++ src/explore-education-statistics-common/src/components/ExportMenu.module.scss
@@ -12,19 +12,22 @@
 
 .exportMenu {
   display: grid;
-  justify-content: start;
+  justify-content: flex-end;
   margin: 0;
   padding: govuk-spacing(2);
-  padding-top: govuk-spacing(0);
+  position: relative;
 
   :global(.govuk-details__text) {
     background: govuk-colour('white');
     border: 2px solid govuk-colour('black');
     border-top: 0;
-    position: absolute;
-    z-index: 10;
     padding-right: govuk-spacing(4);
+    right: 0;
   }
+}
+
+.exportMenu summary {
+  margin-left: auto;
 }
 
 .exportMenu button {
