+++ src/explore-education-statistics-common/src/modules/find-statistics/components/DataBlockTabs.tsx
@@ -12,7 +12,7 @@ import TimePeriodDataTable from '@common/modules/table-tool/components/TimePerio
 import getDefaultTableHeaderConfig from '@common/modules/table-tool/utils/getDefaultTableHeadersConfig';
 import mapTableHeadersConfig from '@common/modules/table-tool/utils/mapTableHeadersConfig';
 import { DataBlock } from '@common/services/types/blocks';
-import React, { ReactNode } from 'react';
+import React, { ReactNode, useRef } from 'react';
 
 const testId = (dataBlock: DataBlock) => `Data block - ${dataBlock.name}`;
 
@@ -51,6 +51,8 @@ const DataBlockTabs = ({
     releaseVersionId,
     getInfographic,
   });
+
+  const dataTableRef = useRef<HTMLElement>(null);
 
   const errorMessage = <WarningMessage>Could not load content</WarningMessage>;
   if (isTableDataError) return errorMessage;
@@ -118,6 +120,7 @@ const DataBlockTabs = ({
                   key={dataBlock.id}
                   captionTitle={dataBlock?.heading}
                   dataBlockId={dataBlock.id}
+                  ref={dataTableRef}
                   footnotesHeadingHiddenText={`for ${dataBlock?.heading}`}
                   fullTable={fullTable}
                   source={dataBlock?.source}
