+++ src/explore-education-statistics-common/src/modules/find-statistics/components/TableExportMenu.tsx
@@ -1,13 +1,11 @@
 import ButtonText from '@common/components/ButtonText';
-import Details from '@common/components/Details';
-import useToggle from '@common/hooks/useToggle';
-import React, { RefObject } from 'react';
 import { Footnote } from '@common/services/types/footnotes';
 import downloadTableOdsFile from '@common/modules/table-tool/components/utils/downloadTableOdsFile';
 import { FullTable } from '@common/modules/table-tool/types/fullTable';
-import styles from './ExportTableButton.module.scss';
+import ExportMenu from '@common/components/ExportMenu';
+import React, { RefObject } from 'react';
 
-interface ExportTableButtonProps {
+interface Props {
   fileName?: string | undefined;
   footnotes?: Footnote[];
   fullTable?: FullTable;
@@ -15,15 +13,13 @@ interface ExportTableButtonProps {
   title?: string | undefined;
 }
 
-const ExportTableButton: React.FC<ExportTableButtonProps> = ({
+export default function ExportTableButton({
   fileName,
   footnotes = [],
   fullTable,
   title,
   tableRef,
-}) => {
-  const [isOpen, toggleOpened] = useToggle(false);
-
+}: Props) {
   const tableFootnotes = fullTable
     ? fullTable.subjectMeta.footnotes
     : footnotes;
@@ -53,31 +49,15 @@ const ExportTableButton: React.FC<ExportTableButtonProps> = ({
   };
 
   return (
-    <Details
-      summary="Export options"
-      hiddenText={`for ${title || 'table'}`}
-      open={isOpen}
-      onToggle={toggleOpened}
-      className={styles.exportMenu}
-    >
-      {isOpen && (
-        <ul
-          className={`${styles.exportMenuList} govuk-!-font-size-16`}
-          role="menu"
-        >
-          <li role="menuitem">
-            <ButtonText onClick={exportToOds}>Download table as ODS</ButtonText>
-          </li>
-          <li role="menuitem">
-            <ButtonText onClick={copyTableToClipboard}>
-              Copy table to clipboard
-            </ButtonText>
-          </li>
-          {/* need to add CSV functionality here but seems we need access to release table query */}
-        </ul>
-      )}
-    </Details>
+    <ExportMenu title={title || 'table'}>
+      <li role="menuitem">
+        <ButtonText onClick={exportToOds}>Download table as ODS</ButtonText>
+      </li>
+      <li role="menuitem">
+        <ButtonText onClick={copyTableToClipboard}>
+          Copy table to clipboard
+        </ButtonText>
+      </li>
+    </ExportMenu>
   );
-};
-
-export default ExportTableButton;
+}
